[GLOBAL][UPDATE] Add man page, add counter for flags

+++ holberton.h
@@ -9,13 +9,13 @@ int _putchar(char c, int *);
 
 int di_format(char *, long int, int *);
 
-int o_format(char *, unsigned int, int *);
+int o_format(char *, unsigned long int, int *);
 
-int u_format(char *, unsigned int, int *);
+int u_format(char *, unsigned long int, int *);
 
 int x_format(char *, unsigned long int, int *);
 
-int X_format(char *, unsigned int, int *);
+int X_format(char *, unsigned long int, int *);
 
 int c_format(char *, char, int *);
